% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_surnames.R
\name{analyze_surnames}
\alias{analyze_surnames}
\title{Categorize names into origins}
\usage{
analyze_surnames(x)
}
\arguments{
\item{x}{a data frame returned by \code{read_honor}
which contains information about one's name, year,
and honor}
}
\value{
a data frame indicating every student's
potential cultural origin
}
\description{
Categorize names into origins
}
