% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cantoneseLastName.R
\docType{data}
\name{cantoneseLastName}
\alias{cantoneseLastName}
\title{list of cantonese last names}
\format{A list with 65 rows}
\usage{
readHonor(students, cantoneseLastName)
}
\description{
A dataset containing all the possible pinyin last names
}
\details{
\itemize{
  \item name
}
}
\keyword{datasets}
