% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chineseLastName.R
\docType{data}
\name{chineseLastName}
\alias{chineseLastName}
\title{list of chinese last names}
\format{A list with 396 rows}
\usage{
readHonor(students, chineseLastName)
}
\description{
A dataset containing all the possible pinyin last names
}
\details{
\itemize{
  \item name
}
}
\keyword{datasets}
