% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/honorStats.R
\name{honorStats}
\alias{honorStats}
\title{create a simple count list of academic honors}
\usage{
honorStats(honorList)
}
\arguments{
\item{honorList}{data frame return by readHonor()}
}
\value{
data frame of \code{year} \code{total} \code{summa}\code{magna}\code{cum}\code{sigmaXi} and \code{phiBetaKappa}
}
\description{
create a simple count list of academic honors
}
\examples{
honorStats(readHonor(DataSummary, chineseLastName) )
}
