% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/japaneseLastName.R
\docType{data}
\name{japaneseLastName}
\alias{japaneseLastName}
\title{list of japanese last names}
\format{A list with 983 rows}
\usage{
readHonor(students, japaneseLastName)
}
\description{
A dataset containing all the possible pinyin last names
}
\details{
\itemize{
  \item name
}
}
\keyword{datasets}
