% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_roster.R
\name{join_roster}
\alias{join_roster}
\title{This function match a student athlete's roster name with his graduation catalog name}
\usage{
join_roster(students, roster)
}
\description{
This function match a student athlete's roster name with his graduation catalog name
}
