% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/originAverage.R
\name{originAverage}
\alias{originAverage}
\title{It allows us to create a graph that illustrates the average percentage of certain honor for all countries of origin}
\usage{
originAverage(honor)
}
\arguments{
\item{honor}{Summa, Magna or Cum}
}
\value{
The average of many countries and total
}
\description{
It allows us to create a graph that illustrates the average percentage of certain honor for all countries of origin
}
\examples{
originAverage("Summa")
add(10, 1)
}
