% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_summary.R
\name{plot_summary}
\alias{plot_summary}
\title{This function offers users various ways to plot their results}
\usage{
plot_summary(x, type)
}
\description{
@param x a dataframe to be plotted
 @param type indicates which type of summary is desired,
 currently there are "count" and "ratio"
}
\details{
@return a plot
 @import ggplot2
 @export
}
