% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratioIndicator.R
\name{ratioIndicator}
\alias{ratioIndicator}
\title{ratioIndicator()}
\usage{
ratioIndicator(asianSummary)
}
\arguments{
\item{data}{a list of student names and graduation year.}
}
\value{
a data frame that contains ratio of magna, summa,
cum, sigmaXi, and phiBetaKappa between Asian and non-Asian
students
}
\description{
ratioIndicator allows users to examine the ratio between percentage
of Latin honors among students of Asian descent and students of non-
Asian descent. A ratio that is larger than 1 inidicates that the hon
-nor is more frequent among the Asian students and a ratio that is
lower than 1 indicates the opposite. A ratio of 1 indicates the
equal percentage of honors.
}
\examples{
ratioIndicator(honorStats(readHonor(students, totalVeri)))
}
