% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read}
\alias{read}
\title{Plot out all the percentage of Latin honors based on one origin for the past 16 years}
\usage{
read(option, verifier)
}
\arguments{
\item{option}{latin Honor: Summa, Magna, or Cum}

\item{verifier}{chinese,cantonese, or japanese origin}
}
\value{
the percentage of certain latin honor during the past 16 years
}
\description{
Plot out all the percentage of Latin honors based on one origin for the past 16 years
}
\examples{
read ("Summa", chineseLastNames)

}
