% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_honor.R
\name{read_honor}
\alias{read_honor}
\title{This method reads honors such as Summa Cum Laude, Magna Cum Laude, and Phi Beta
Kappa.}
\usage{
read_honor(x)
}
\arguments{
\item{x}{a dataframe including rows \code{name} and others}
}
\value{
a dataframe including rows \code{name} \code{phi_beta_kappa} \code{sigma_xi}
\code{summa} \code{magna} \code{cum} and others
}
\description{
This method reads honors such as Summa Cum Laude, Magna Cum Laude, and Phi Beta
Kappa.
}
