% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readHonor.R
\name{readHonor}
\alias{readHonor}
\title{Pick out last names based on reference list, and introduce academic honors}
\usage{
readHonor(data, verifier)
}
\arguments{
\item{data}{list of graduation names and honors}

\item{verifier}{A list of last names}
}
\value{
data frame of \code{ethnic last names} and \code{honors}
}
\description{
Pick out last names based on reference list, and introduce academic honors
}
\examples{
readHonor(DataSummary, chineseLastName)
}
