% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsSum.R
\name{statsSum}
\alias{statsSum}
\title{statsSum()}
\usage{
statsSum(asianSummary)
}
\arguments{
\item{data}{a list of student names and graduation year.}
}
\value{
a data frame including a column of honors and a column of ratios
}
\description{
Going through the list of summary given by honorStats and totalStats and find
the total sum of all honors of all students. With the total sum for both Asians and non-Asians, we can then calculate
the ratio of the percentage of getting honors
}
\examples{
statsSum(honorStats(readHonor(students, totalVeri)))
}
