% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/students.R
\docType{data}
\name{students}
\alias{students}
\title{names and years of 7874 students}
\format{A data frame with 7874 rows and 2 variables}
\usage{
readHonor(students)
}
\description{
A dataset containing the name of the students and their years
}
\details{
\itemize{
  \item student
  \item year
}
}
\keyword{datasets}
