% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_ratio.R
\name{summarize_ratio}
\alias{summarize_ratio}
\title{This function summarizes the ratio of honor among each ethnic group}
\usage{
summarize_ratio(x)
}
\arguments{
\item{x}{A dataset containing name, honor, origin}
}
\value{
honor ratio for each origin group
}
\description{
This function summarizes the ratio of honor among each ethnic group
}
