% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totalStats.R
\name{totalStats}
\alias{totalStats}
\title{totalStats()}
\usage{
totalStats(data)
}
\arguments{
\item{data}{a list of student names and graduation year.}
}
\value{
a data frame including year, total, summa, magna, cum, sigmaXi, and phiBetaKappa
}
\description{
going through the list of all students and figure out the number of total students, summa, magna, cum, sigmaXi, and phiBetaKappa each year.
}
\examples{
totalStats(students)
}
