% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totalVeri.R
\docType{data}
\name{totalVeri}
\alias{totalVeri}
\title{list of total last names}
\format{A list with 1444 rows}
\usage{
readHonor(students, totalVeri)
}
\description{
A dataset containing all the possible pinyin last names
}
\details{
\itemize{
  \item name
}
}
\keyword{datasets}
