% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_honor.R
\name{read_honor}
\alias{read_honor}
\title{This function process the student info data frame and mutate their academic honors}
\usage{
read_honor(x)
}
\arguments{
\item{x}{the dataframe containing all student info}
}
\value{
a dataframe now containing columns \code{name}, \code{year}, \code{honor},
and \code{value}
}
\description{
We determine whether a student is part of phi beta kappa or sigma xi by detecting
if the name contains "*" or "+"
}
\details{
We also attain a list of indices from rows saying "Bachelor of Arts, Summa Cum Laude"
and find out their places in the data frame, and mutate the data frame based on where
an entry belongs in the list of indices. Then apply %% 4.
}
