% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statsummary.R
\name{statsummary}
\alias{statsummary}
\title{provides different ways of evaluating the given data set}
\usage{
statsummary(type)
}
\arguments{
\item{type}{types of analytical representations, including: summay by count, summary by origin, breakdown, histogram, summary by percentage}
}
\value{
ggplot of graphs
}
\description{
provides different ways of evaluating the given data set
}
\examples{
statsummary("breakdown")
}
